Snapshot tests (#4)
## Why is this change necessary?
The XML files are large with many components. Generation of full
programs is well suited to snapshot testing.


 ## How does this change address the issue?
Adds the syrupy library and implements snapshot testing of a few
programs.


 ## What side effects does this change have?
None


 ## How is this change tested?
Snapshot tests


 ## Other
Dynamically pulling the TipID from the `Tip`, rather than hardcoding it.
Adjusted the LibraryComponent class to be frozen so that more fixtures
can be made in the future using them. This required refactoring the
xml_root attribute, but it wasn't being used as a cached attribute
anyway, that will need to be implemented later.

---------

Co-authored-by: Eli Fine <ejfine@gmail.com>