DRY up steps (#17)
## Why is this change necessary?
There was some duplication in the generation of the steps.


 ## How does this change address the issue?
Creates some pydantic models for some of the JSON representation within
the steps.

Initializes models for Well and Deck Section at the beginning of
add_values, and then uses them everywhere in the XML generation.


 ## What side effects does this change have?
Cleaner code.


 ## How is this change tested?
Existing snapshot tests.

---------

Co-authored-by: Eli Fine <ejfine@gmail.com>