Delist py313 compatiblity (#19)
## Why is this change necessary?
A misconfiguration in CI led to thinking that this would work in Python
3.13. But in fact, `lxml` library does not yet support 3.13


 ## How does this change address the issue?
Updates CI to fix the underlying issue.  Drops Python 3.13 support.


 ## What side effects does this change have?
None


 ## How is this change tested?
CI


 ## Other
Added building documentation as part of CI

Fixed the Readme badge so that it only shows the status on `main`...not
some random failure if a feature branch had a problem