Create API for the MultiDispense step (#28)
## Why is this change necessary?
It offers much more control over liquid transfers than the simple
Transfer step


 ## How does this change address the issue?
Adds the ability to use MultiDispense. Not all features are
parametrized...this was largely to enable an IMCS workflow.
Specifically, it only actually does a single dispense, and it does not
yet support mixing.


 ## What side effects does this change have?
None


 ## How is this change tested?
Snapshot tests and simulating those in ViaLab


 ## Other
Ran Ruff on the Jupyter notebooks to clean them up

Rather than needing an enumeration of all possible deck positions, the
labware footprint is now dynamically accessed to find the appropriate
deck position

Supporting Reservoirs and Tubes required an abstraction of Labware,
which necessitated renaming some things

Found out that the Reservoir Integra definition is a "version=2", so
needed to setup a dictionary to map those content types.

There's some code duplication that pylint detects between the Transfer
and MultiDispense steps. A lot of it was refactored to DRY it up, but
some remains that seems less amenable. Keep an eye on it going forward
to ensure it stays reasonably DRY.

---------

Co-authored-by: Eli Fine <ejfine@gmail.com>