(bugfix) Portrait Plate Orientation (#29)
## Why is this change necessary?
When a plate was in portrait orientation in the B Position, the ASSIST
Plus did not correctly pipette into it.


 ## How does this change address the issue?
When searching for the correct deck section, length and width are
swapped if the labware is in portrait mode.


 ## What side effects does this change have?
None


 ## How is this change tested?
Unit and snapshot tests, and observing in the ViaLab UI that a DWP in
Portrait mode in position B renders at the correct size (it was
previously rendering much too small)...and also the previous snapshot of
15 mL Tube holder is now correct size (was much smaller before)


 ## Other
Added start heights to the IMCS demo and reduced mixes when in debug
mode